requirements.txt
attrs
python-dotenv
tgalice[server]
mongomock
pymongo
dnspython
regex
requests
timezonefinder
pytz
qrcode
